+++ intelmq/tests/bots/collectors/bitsight/test_collector.py
@@ -1,6 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-Testing Bitsight Ciberfeed Stream
-"""
-
-import intelmq.bots.collectors.bitsight.collector